refactor(github): consolidate type definitions and remove redundant code
- Merge batchRepoTypes.ts and types.ts into a single file
- Move RepoData type import from services/github to types/github
- Remove duplicate type definitions across components
- Delete unused development tools components
- Update import paths for hooks and other utilities
