+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 // import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import { RouterProvider, type createRouter } from "@tanstack/react-router";
-import type { FunctionComponent } from "./common/types";
-// import { TanStackRouterDevelopmentTools } from "./components/utils/development-tools/TanStackRouterDevelopmentTools";
+import type { FunctionComponent } from "./types";
+// import { TanStackRouterDevTools } from "./dev-tools";
 
 const queryClient = new QueryClient();
 
@@ -12,7 +12,7 @@ const App = ({ router }: AppProps): FunctionComponent => {
 	return (
 		<QueryClientProvider client={queryClient}>
 			<RouterProvider router={router} />
-			{/* <TanStackRouterDevelopmentTools
+			{/* <TanStackRouterDevTools
 				router={router}
 				initialIsOpen={false}
 				position="bottom-right"
