--- src/common/utils.ts
@@ -1,1 +1,0 @@
-export const isProduction = import.meta.env.MODE === "production";
