+++ src/components/github/analysis/tabs/CommitsTab.tsx
@@ -5,13 +5,13 @@ import {
 	BarChart as BarChartIcon,
 	Timeline as TimelineIcon,
 } from "@mui/icons-material";
-import type { RepoData } from "../../types/types.ts";
 import UserTabItem from "../components/UserTabItem";
-import AnalysisTabLayout from "../components/AnalysisTabLayout";
 import TabDataTable from "../components/TabDataTable";
 import CommitBarChart from "../components/CommitBarChart";
 import { commitsTheme } from "../components/AnalysisThemes";
 import AllContributorsCommitChart from "../components/AllContributorsCommitChart";
+import { RepoData } from "../../../../services/github";
+import AnalysisTabLayout from "../components/layout/AnalysisTabLayout.tsx";
 
 interface CommitsTabProps {
 	data: RepoData;
