+++ src/components/github/analysis/tabs/RepoTabs.tsx
@@ -1,8 +1,7 @@
 import { Box, Grow } from "@mui/material";
-import type { FunctionComponent } from "../../../../common/types.ts";
 import "../../../../styles/FormStyles.css";
 
-const RepoTabs = (): FunctionComponent => {
+const RepoTabs = (): JSX.Element => {
 	return (
 		<Grow in timeout={700}>
 			<Box className="tab-container">
