+++ src/components/github/analysis/tabs/SummaryTab.tsx
@@ -12,7 +12,8 @@ import {
 	Avatar,
 	LinearProgress,
 } from "@mui/material";
-import type { RepoData, ContributorStats } from "../../types/types.ts";
+import { RepoData } from "../../../../services/github";
+import { ContributorStats } from "../../../../types";
 
 interface SummaryTabProps {
 	data: RepoData;
