+++ src/dev-tools/ReactHookFormDevTools.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import { isProduction } from "../../../common/utils";
+import { isProduction } from "../utils";
 
-export const ReactHookFormDevelopmentTools = isProduction
+export const ReactHookFormDevTools = isProduction
 	? (): null => null
 	: React.lazy(() =>
 			import("@hookform/devtools").then((result) => ({
 				default: result.DevTool,
 			}))
-	  );
+	  );
