+++ src/dev-tools/TanStackRouterDevTools.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import { isProduction } from "../../../common/utils";
+import { isProduction } from "../utils";
 
-export const TanStackRouterDevelopmentTools = isProduction
+export const TanStackRouterDevTools = isProduction
 	? (): null => null
 	: React.lazy(() =>
 			import("@tanstack/router-devtools").then((result) => ({
 				default: result.TanStackRouterDevtools,
 			}))
-	  );
+	  );
