+++ src/hooks/useGitHubToken.ts
@@ -88,4 +88,4 @@ export const useGitHubToken = (): {
 		deleteToken,
 		handleTokenMessageClose,
 	};
-};
+};
