+++ src/utils/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Common utility functions and constants
+ */
+
+export const isProduction = import.meta.env.MODE === "production";
